control panel fixes (rip Pikachu)
